web/code/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
Flask==2.0.3
gTTS==2.2.3
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
packaging==21.3
Pillow==9.0.1
playsound==1.3.0
pyparsing==3.0.7
pytesseract==0.3.9
requests==2.27.1
six==1.16.0
urllib3==1.26.8
Werkzeug==2.0.3
